% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_data.R
\docType{data}
\name{comorbidity_data}
\alias{comorbidity_data}
\title{Sampele comorbidity data set}
\format{100 rows and 10 column matrix}
\usage{
comorbidity_data
}
\description{
This is the sample data set for Package PDN
}
\keyword{dataset}
