% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_data.R
\docType{data}
\name{demographic_data}
\alias{demographic_data}
\title{Sampele demographic data set}
\format{100 rows and 5 column matrix}
\usage{
demographic_data
}
\description{
This is one of the sample data set.
It contains the demographic information of each patients
It has five variables which are sex, race, hispan, dshyr and prime
}
\keyword{dataset}
