\name{PDN-package}
\alias{PDN-package}
\docType{package}
\title{
Personized Disease Network
}
\description{
  Building patient level networks for prediction of
  medical outcomes and draw the cluster of network.
  This pakcage is based on paper Personalized disease
  networks (PDN) for understanding and predicting cardiovascular
  diseases and other complex processes.
}
\details{
\tabular{ll}{
Package: \tab PDN\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-08-14\cr
License: \tab GPL-2\cr
}
}
\author{Javier Cabrera and Fei Wang\cr
Maintainer: Zhenbang Wang <bruce1edward@gmail.com>
}
\references{Cabrera, Javier and Kostis, John B and Wang, Fei and Kostis, William J. (2016)
  \emph{Personalized Disease Networks in Predicting Cardiovascular Outcomes}}
\keyword{models}
\keyword{regression}
\keyword{package}

