% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_data.R
\docType{data}
\name{survival_data}
\alias{survival_data}
\title{Sampele Survival data set}
\format{100 rows and 2 column matrix}
\usage{
survival_data
}
\description{
This is the sample data set for function datecut.
The first column is the time difference between the patient admission date and death date
The second column is the indictor variable with 1 corresponding to patient death, 0 corresponding to patient alive
}
\keyword{dataset}
